Update lang_french.php
line 11 cannot match line 79 and 80. "the" can be both singular and plural in english but there is no perfect match for it in French. Plural is needed for line 79 ("J'accepte les") and singular is needed for line 80 ("J'accepte la"). I chose to not use any of them to avoid any confusion and went with "J'accepte"("I agree") instead.